Map Model Adjustments
The MapModel class was updated and partially-rewritten to account for
the off loading of the map to the ATWorldDB class. The class's
file-reading methods were updated so that game entity information could
be sent to the database class for storage. In addition, the class;s
constructor and map-loading order were organized to account for both
missing and already existing game databases. The modifiable game
introduction functionality was removed alongside some of the other
non-working and unused Mini2 functions.