Rooms.txt Symbol Adjustments/Replacements
In an attempt to fix some of ATWorldDB's database-populating methods,
some of the text in Rooms.txt was altered to filter out special symbols
like :, '. and ". While ATWorldDB's methods were explored and found to
have issues with missing single quotations, the changes here are being
pushed in order to ensure unintended errors.