Controller Package Save Methods
The included classes include placeholder methods that will be used to
send data to the MapModel class for saving. These classes also witnessed
the removal of whitespace, empty constructors, and the addition of an
InventoryOperations class. This class mainly serves as a name space as
the connected behavior of player inventory saving and room item saving
requires the use of additional methods that don't fall in either Player
or Room name-wise.