GameController Room Setting and Current Player Maintenance
The GameController class was modified to include the setCurrentRoom()
method so that the MapModel could set the current player's room to the
default after creating a new save profile or setting it to an existing
room when loading a save profile. The class was also modified to include
a Player field that is used to maintain's the game's current playing
character.